feat: deps, sources, repos

+++ app/src/main/java/ru/sicampus/bootcamp2026/data/dto/UserDto.kt
@@ -2,14 +2,12 @@ package ru.sicampus.bootcamp2026.data.dto
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
-import ru.sicampus.bootcamp2026.domain.models.User
-import java.time.Instant
 
 
 @Serializable
 data class UserDto(
     @SerialName("id")
-    val id: Int?,
+    val id: Long?,
     @SerialName("email")
     val email: String?,
     @SerialName("firstName")
